fix(review): address MEDIUM/LOW findings from animated-light weight-maps review
- MEDIUM #1: use checked_mul for total_probes and probe_stride multiplication
  in ShVolumeSection::from_bytes to prevent silent overflow on crafted PRL data.
- MEDIUM #2: add (entry.offset + entry.count) <= texel_lights.len() bounds
  check in AnimatedLightWeightMapsSection::is_consistent, using checked_add
  to guard against overflow before the cast.
- MEDIUM #4: add comment on the texel_idx u32 addition in
  animated_lightmap_compose.wgsl noting it is safe within validated PRL budgets.
- LOW #1: fix set_active out-of-range check to compare in the wider usize type
  (slot >= animated_light_count as usize) instead of truncating slot as u32.
- LOW #2: promote debug_assert! to assert! in create_storage_buffer so a
  future empty-bytes regression is caught in release builds.

+++ postretro/src/render/sh_volume.rs
@@ -116,7 +116,7 @@ impl AnimatedLightBuffers {
     /// infallible because the scripting wiring isn't in place yet; promoting
     /// the signature to `Result` is a scripting-era decision.
     pub fn set_active(&mut self, slot: usize, active: bool) {
-        if (slot as u32) >= self.animated_light_count {
+        if slot >= self.animated_light_count as usize {
             if !self.oor_warned {
                 self.oor_warned = true;
                 log::warn!(
